Development Commit
This change sets up the initial project directory structure. The domain
directory has not been added since there are initially no entities to add.

This baseline commit provides an entry point to the program in main.py, which
when executed creates a main window and a controller instance. The controller
instance is passed a reference to the main window.