Create application window & video, table widgets
This patch sets up the majority of the project structure and adds a
starting point for the development of the project.

A menu bar has been added to the top of the window of the main
application. A file sub-menu exists in the main menu bar, and in it an
option which allows the user to open a file browser to select a local
video file to load when clicked.

On the left half of the application is a video streaming widget. This
widget will be able to stream the selected video for the user. Since
there does not yet exist a play button, the video plays as soon as its
selected.

On the right half of the applicatoin is a custom encoding table widget,
and buttons to add rows and columns to it.

Testing Steps
  1. Run main.py to start the application.
  2. Verify that there exists a menu bar at the top of the application.
  3. Verify that there exists a file sub-menu at the top of the
     application.
  4. Verify that the main application evenly divides its horizontal
     layout to the video and table widgets.
  5. Open the file browser by clicking on File -> Load video
  6. Select a MP4 testing video file
  7. Verify that the video is streamed in the video widget.
  8. Verify that the add row button adds a row to the table.
  9. Verify that the add column button adds a column to the table.

Type: New Feature