Add "All supported types" default to file browser
This patch adds a new file type to the file dialog window that opens
when the user loads a local video file. The new file type is an "all
supported file formats" option which includes all supported mime types
on the user's system. This option is made to be the default option.

Before this patch, a warning message would appear when certain
directories were switched to using the macOS native file browser. To
rememedy this, the QT-specific file browser is used instead.

Testing Steps
  1. Run main.py to start the application.
  2. Click the load video option in the menu bar.
  3. Verify that an option to select "All supported formats" appears and
     is the default option.
  4. Verify that a video can be run with no errors.
  5. Verify that switching directories does not cause an exception to be
      thrown.

Type: Bug Fix
Type: New Feature