Add ability to export table data to CSV
This patch gives the user the ability to save table data (rows and columns)
to CSV file format. It adds features like a new sub-menu to allow for the
user to bring up their native file browser for saving. When saving it converts
all the data within the table panel to csv form and then stores as a byte
array.

Testing Steps
1. Run main.py to start the application.
2. Add dummy data to any slots within the table.
3. Click the export option in the menu bar.
4. Click the save to save table data button within export.
4. Verify that a file browser pops up and you can save the data.
5. Verify how the data is formatted within the location you saved to.

Type: New Feature