Adds controls for playback speed
This patch adds a playback speed combobox located in the media control panel,
where the user can select playback speeds of 0.25, 0.5, 0.75, 1.0, 1.25, 1.5,
1.75, and 2.0. The combobox itself is a custom combobox in order to print a
label in the combobox text.

The playback speed is not recognized if it is changed before the video is
loaded and played. This "bug" may be fixed with the addition of the play and
pause button. If not, it will need to be addressed in the future.

Testing Steps
  1. Run main.py to start the application.
  2. Load a video and play it.
  3. Verify that the playback speed changes accurately according each option in
     the playback speed combobox.

Type: New Feature