Adds timestamp feature to media control panel
This patch adds a new feature within the media control panel. It displays the
current time out of the entire time for the media file that is playing. The
format is displayed in: hours:minutes:seconds for both the current time
and the entire time. It uses the position changed signal to calculate the time.

Testing Steps
  1. Run main.py to start the application.
  2. Click the load video option in the menu bar.
  3. Select a media file to play
  4. Verify that a video can be run with no errors.
  5. Verify that the time stamp is being updated as the video progresses.

Type: New Feature