Adds a title bar to the table panel
Adds a title bar to the tabel panel centered above the encoding table headers.
The title bar is editable by the user, and its width resizes according to the
width of its text content. The default value of the title is "Table Title".

This patch also adds a pop-up input dialog box which appears on the
startup for new sessions. The pop-up requests an initial table name from
the user.

Testing Steps
  1. Run then application
  2. Start a new session
  3. Verify that a pop-up input dialog box appears, requesting an
     initial title for the table
  4. Verify that the default table title name is "Table Title"
  5. Enter "Test" in the input dialog box
  6. Verify that the table title updated, and its width is resized
     appropriately
  7. Verify that the title can be edited, and that the width of the label
     resizes according to the title content. The width should always fit the
     title content plus 25px of padding
  8. Restart the application, this time loading a previously saved
     session
  9. Verify that there is no pop-up upon startup

Type: New Feature