Dev add buttons to button panel (#19)
Adds a change to the Coding Assistance Panel where the "+" button now opens a dialog that allows the user add a new button to the Coding Assistance Panel. The dialog allows the user to label the button and assign a hotkey to the button. Hotkeys cannot be reused.

Testing Steps
  1. Open the Application and start a new session
  2. Click the + button under "Coding Assistance Buttons"
  3. Give a title and a hotkey "a" to the button and press the "Create" button
  4. Verify the button has been added and has the correct label
  5. Press the "Create" button again without changing the parameters and verify the text "This hotkey is already being used!" shows in the dialog
  6. Change the hotkey to "s" and click the "Create" button
  7. Verify a new button is created.

Type: New Feature