Updates the UI + functionality of welcome page
This patch updates the UI for the welcome page in order to provide a
more user-friendly welcome page with additional functionality.

Before this change, the user interface design for the welcome page was a
singular window containing a list of buttons for creating and loading a
session. With this change, the welcome page is a stacked widget,
allowing the user to manage their sessions on the default page, or
create a new session on the "Create New Session" page.

The user can now delete a specified session, by clicking the "X" icon
next to a session. Additionally, a user can clear all loaded sessions
with the "Clear all sessions" button.

The user can view their saved sessions in a scrollable viewport. This
allows the display of sessions to cover all sessions, regardless of
size.

The user can specify a unique session name on the session creation page.
Before this patch, the session name was generated from the timestamp of
its first save. Now, the user can specify their own label.

Testing Steps:
  1. Run the application
  2. Click the "Clear all sessions" button to remove all previously
     saved sessions.
  3. On the welcome page, click the "Create New Session" button under
     the "Getting Started" section.
  4. Enter a unique session name, "Session A" and click "Ok".
  5. Fill in dummy data into the encoding table and exit the
     application.
  5. Repeat steps 1., 3., 4., and 5. to create new sessions "Session B"
     and "Session C".
  6. Run the application
  7. Click on "Session A" under the "Open Recent" section and verify the
     dummy data entered for the session is preserved.
  8. Repeat step 7 for "Session B" and "Session C".
  9. Run the application
  10. Click on the "x" icon next to "Session B" and confirm that it is
      removed from the GUI and "Session C" moved upwards in the list.
  11. Restart the application and verify that "Session B" no longer
      exists
  12. Click on the "Clear all sessions" button and verify that all
      sessions are removed from the GUI
  13. Restart the application and verify that the session list is empty
  14. Create a new session, "Session A"
  15. Restart the application and create a new session called "Session
      A", and verify that the session is not created
  16. Verify that all functionality of a created/loaded session function
      properly.

Type: New Feature