Dev set default csv file name as table name (#27)
Sets default csv export name as table name

This patch makes it so the default name is set to the table title. If the user
has not given the table a name it will default to the name Table Title. This can
be changed in the future if we want to set the default name to something other
then that. Additionally, if the name is left blank it sets the name as
your_table_data.csv. This patch was done to add ease of function for the user
so they would be familiar with the name of the file they saved if did not change
its default name.

  1. Run the program
  2. Don't edit the table name
  3. Export the table data
  4. Check that the csv file name is Table Title
  5. Close the program
  6. Run again repeating same steps this time name table
  7. Check that when you export the csv file name it is the same
      as the table title
  8. Close the program
  9. Run again this time deleting table name so it is empty string
  10. Check that when you export the csv file name it is the same
        as the table title