Dev session entity and session manager (#28)
Change the architecture to follow the MVC architecture

In this patch we removed all our logic that allowed session persistence from the view
and moved it to a session manager. The controller now connects on a signal when the
window closes to either grab the data from the view and set it to the manager, or grab
data from the manager and load it into the view. We also added one more component,
a session entity, which acts a container that holds the most recent session. This patch
was done to allow for more modularity within our codebase as well as ease in future
updates like loading a session while the program is running.

  1. Run the program
  2. Create a new session (Session A)
  3. Add dummy data and change table name
  4. Close the session
  5. Verify upon loading this session everything persists
  6. Repeats this process with 2 more session (Session B
      and Session C)
  7. Delete Session B
  8. Verify that did not change Session A or C
  9. Create a new Session B and fill with data
  10. Verify that the new Session B you made is loaded
        correctly

Type: Refactor