Fixes lowercase session id bug
Before this change, all session ID values written through QSettings were
titleized, using python's title() method. This prevented the user from
creating a session with a lowercase session ID, as it would not equal
its titleized counterpart. This patch removes the title() method,
allowing the session ID to be written as is, and consequently allows for
session IDs of all cases.

This patch also updates the check for session ID uniqueness to be
case-insensitive. This change is to support all platforms, as the
Windows registry uses case-insensitive keys.

Testing Steps:
  1. Run the application and clear all sessions
  2. Create a new session and give it an ID="test"
  3. Write dummy data in the encoding table
  4. Restart the application and verify that the session in the welcome
     screen is labeled "test" and not "Test"
  5. Load the "test" session and verify the dummy data has persisted
  6. Restart the application
  7. Attempt to create a new session called "Test", and verify it is not
     allowed
  8. Repeat steps 1-5, but with a new session ID="Test"

Type: Bug Fix