Adds a delete row and delete column button
In this patch, a user can select a row or column and then press the "delete
selected row" or "delete selected column" button to delete the selected
row/column. The buttons also work if a user selects a specific cell
instead of a whole row/column. The first row and first column cannot
be deleted currently.

Testing Steps:
  1. Run main.py to start the application.
  2. Select a column (not the first) and press the delete column button, make
     sure column is deleted.
  3. Select the first column, press the delete button, and make sure it is
     NOT deleted.
  4. Select a row (not the first) and press the delete row button make sure
     row is deleted.
  5. Verify that the last row of the table cannot be deleted.
  6. Select a column and click the delete row button, make sure nothing is
     deleted.
  7. Delete all rows but one, then select the only row left and click the
     delete row button, make sure it is NOT deleted.

Type: New Feature