Adds session persistence to encoding buttons
This patch adds session persistence to encoding buttons. This means that
all encoding buttons in the button panel are saved at the session level.
When the session is loaded again, the buttons are automatically loaded
with their assigned hotkeys.

This patch also adds back support for encoding button removal. If a
button is removed, and the session is exited, then the button does not
persist.

Additionally, this patch removes hotkeys from global settings. Buttons
loaded at the global settings level do not have assigned hotkeys, as
hotkeys are a session level piece of data. The functionality to assign
hotkeys to loaded global button definitions should be added in a future
patch.

Testing Steps:
  1. Run the application and clear all sessions
  2. Create a new session, "Session A"
  3. Create three buttons, "button 1", "button 2", and "button 3",
     with hotkeys "a", "b", "c", and dummy data. Save the buttons
     globally
  4. Verify that the button functionality works as expected
  5. Repeat steps 2-4, but with "Session B", and "button 4",
     "button 5", and "button 6", dummy data, and the same hotkeys
  6. Load "Session A"
  7. Verify that buttons 1-3 are loaded, and the functionality
     works as expected
  8. Repeat steps 6-7, but with "Session B"
  9. Create a new session, "Session C"
  10. Load "button 1"
  11. Reload "Session C" and verify the button persists and
      functionality works as expected
  12. Load "Session A"
  13. Remove "button 2" and verify its removed from the panel
  14. Verify the hotkey "b" does not do anything
  15. Reload "Session A" and verify "button 2" is gone

Type: New Feature