Adds second iteration scalable scrubbing bar
Before this patch, the application used a singular scrubbing bar. The
scrubbing bar represented the state of the multimedia playback, and
could be manipulated to change the associated media state. This patch
adds two new slider bars, that when combined, provide functionality to
scale the scrubbing bar. The top bar is a scaling bar, and has two
handles. The range between the handles represents the range of the
scrubber bar, and as a result will zoom the scrubber bar to reflect its
range. The second bar is a progress bar, and is disabled by default.
This bar is useful for the user to see where the scrubber bar head is at
all times, especially if it's outside the range set by the scaling bar.

This patch also adds tick intervals to the scrubber bar. The default
value is 100, but will adjust to a lower number as the zoom level
increases. In the future, this value should correspond closely to the
frames per second of the video.

This patch uses the superqt module. See integration steps before
testing.

Testing Steps:
  1. Run the application
  2. Verify the defaults of the scalable scrubbing bar
     2a. The scaling bar defaults to a range and value of [0,100]
     2b. The progress bar (middle bar) is disabled
     2c. The scrubber bar is enabled
  3. Load a local video file
  4. Verify that the scaling bar range and value is [0, duration of
     video]
  5. Verify that the scrubber bar works as expected, and that the
     progress bar updates synchronously with the scrubber bar.
  6. Restart the video and pause it at the beginning
  7. Set the scaling bar to include the first half of the video duration
  8. Verify that the scrubber bar has zoomed in
  9. Play the video and verify that the head in the scrubber bar exits
     the slider range at the halfway point of the video
  10. Verify that the progress bar range has not changed, and is still
      visible as the head traverses the second half of the video
  11. Adjust the scaling bar to "catch up" with the progress bar value
  12. Verify the functionality of step 11

Integration Steps:
  1. Verify that your virtual environment is activated
  2. Install the superqt module by following Step 6 in the README
     instructions

Type: New Feature