Adds tick marks to the scaling and progress bars
This patch adds tick marks below the scaling and progress bar views of
the scalable scrubbing bar. The tick marks are placed below the sliders,
and have a default interval of 100. In the future, once the video fps is
known, the default value may change accordingly.

Testing Steps:
  1. Run the application
  2. Create or load a session
  3. Load a video
  4. Verify that there are tick marks below all three sliders that make
     up the scalable scrubbing bar

Type: New Feature