Sets the tick interval according to video fps
This patch adds functionality, with the utilization of the opencv
module, to obtain the frames per second of any loaded video. This patch
gets the frames per second of the loaded video and sets the tick marks
of the scaling bars according to the interval between frames.

Testing Steps:
  1. Add a print statement to the on_video_duration_changed slot in the
     WindowController class to reveal the frames per second data member
     once a video is loaded
  2. Run the application
  3. Create or load a session
  4. Load a video file
  5. Verify that the printed frames per second is correct
  6. Verify that the tick mark interval of the scaling and progress bars
     are set to the frames per millisecond equivalent value
  7. Zoom in, using the scaling bar, past 50% and verify that the tick
     mark value increment is divided in half.

Integration Steps:
  1. Use pip to install the opencv-python module
     https://pypi.org/project/opencv-python/

Type: New Feature