Changes the scaling bar labels to timestamps
This patch changes the label values from millisecond values to timestamp
strings. In order to accomplish this, a custom labeled range slider was
implemented, which paints a custom timestamp label above each handle.

With the current implementation, the labels will slide over when they
get too close together. However, this patch does not implement a fix for
this behavior when the labels are close to the slider minimum or maximum
horizontal positions.

This patch also adds some logic to deal with out of bounds timestamps,
but the values are hard-coded and do not work for all screen sizes. If
it doesn't work on a screen size, then the timestamp is simply cut off.

Testing Steps:
  1. Run the application
  2. Create or load a session
  3. Load a video file
  4. Verify that the scaling bar has timestamp labels above each handle.
  5. Move the left and right handles around and verify that the
     timestamp values are accurate to the progress bar.
  6. Verify that the labels do not overlap when the handles are in the
     middle of the scaling bar.

Type: New Feature