Removes python-opencv dependency
This patch removes the python-opencv dependency from the README
instructions, and removes all opencv related code from the codebase. The
opencv module was used to get the frames per second of the loaded video.
This information is no longer necessary as the tick intervals are set
independent from the frame rate. In addition, we have come across
installation errors caused from the python-opencv dependency. For this
reason, we are removing it from the application.

This patch not only removes the frames per second data retrieval, but it
also removes all QSlider tick marks as well. The slider tick intervals
were based on the video frame rate found from opencv. The QSlider tick
marks will not be used in the future of this application, as they are
being replaced by a custom tick mark bar.

Testing Instructions:
  1. Run the application and verify all functionality works as expected.

Type: Bug Fix