Adds functionality that sets encoding table title based on the table title input field from the session creation window (#44)
Before this patch, there was a table title input field on the "Create New Session" window that had no functionality. This patch adds functionality so that if a user enters a table title, that title will be displayed above the table when the main application window launches. If a user creates a new session without specifying a table title, the default title will be "Table Title".

Testing steps:

Run the application
Clear all sessions
Create a new session "Session A" and enter "Session A" for the table title as well
Check that the table title is "Session A"
Close the application
Reopen Session A, and check that the table title is still "Session A"
Edit the table title to "Session 1"
Close the application and reopen Session A, and check that the table title is still "Session 1"
Close the application and create a new session "Session B" and enter "Session B" for the table title
Check that the table title is "Session B"
Close the application and reopen Session A, and make sure the table title is still "Session 1"

Type: New Feature