Merge `feat_settings_from_welcome` into `main`
* Adds the ability to access user settings from the welcome page

The changes made in this patch was the creation of a user settings controller
to allow the user to access user settings from the welcome page. Many of the
features added were just refactors of our old code base in order to make that
code work with the implementation of the user settings controller. This patch
was done as it made sense in the users work flow to be able to access settings
from the welcome page since they applied to every session populated in the
application.

Testing Steps:
  1. Load the application
  2. Verify clicking the user settings button brings up the user settings prompt
  3. Close the application and verify no errors
  4. Reopen the application and user settings and adjust all values for the table
  5. Create a session and verify the table settings were applied
  6. Create a button called button1 in the application
  7. Create a button called button2 in the application
  8. Verify they exist in user settings by opening user settings inside the current
      session
  9. Open user settings within the session and edit the table settings, verifying
      that the table was updated
  10. Close the session
  11. Reopen and verify the exist in user setting accessed from the welcome page
  12. Delete button1 from user settings
  13. Verify the change was made in user settings accessed from inside a session
        and from the welcome page

Type: New Feature

* Close Dialogs Automatically + Edit Button Definition List in Real Time

This patch allows five dialogs to close automatically when the user inputs their selections. These dialogs being the add button dialog, the delete button dialog, the remove button dialog, the edit button dialog, and the create button dialog. Additionally, this patch allows the button definition list within the user settings dialog to automatically update as soon as a button definition is edited. Finally, this patch adds to the README by telling the user that when creating button definitions, button definition names must be globally unique.

Testing Steps:
  1. Create a new session
  2. Click "+" in the "Coding Assistance Buttons" panel
  3. Create a new button and click "Yes" when prompted to save the button definition
  4. Click "Load Button"
  5. Click "Load Button" and ensure the dialog closes automatically
  6. Close the dialog
  7. Click "-" in the "Coding Assistance Buttons" panel
  8. Click "Delete" and ensure the dialog closes automatically
  9. Click "Settings" in the tool bar and click "Settings"
  10. Click "Edit Definition"
  11. Enter the name of the button created in step 2 and click "Edit"
  12. Give the button a new name and click "Edit"
  13. Ensure both dialogs have closed automatically
  14. Ensure the button definition edited has populated in the list automatically
  15. Click "Remove Definition"
  16. Enter the name of the edited button definition and click "Remove Button Definition"
  17. Ensure the dialog closes automatically
  18. Click "Remove Definition"
  19. Click "Clear all Button Definitions" and click "Yes" when prompted
  20. Ensure the dialogs close automatically
  21. Repeat steps 18-20 and click "No" when prompted

Type: New Feature