Merge scalable scrubbing bar branch into `main`
* Adds labeled scalable tick marks to sliders

This patch adds a custom scalable tick-mark bar to each slider in the
scalable scrubbing bar. The tick-mark bars make assumptions about the
slider width, margins, and positioning. In the future, it would be
reasonable to package the slider and tick mark bars into one container
class, in order to get rid of the required assumptions.

Testing Steps:
  1. Run the application
  2. Verify that there are no visible tick-marks under the bars
  3. Load a video
  4. Verify that the tick-mark bars are now present
  5. Verify the tick-mark bars of the scaling and progress bars.
     5a. The ticks should start at timestamp 0, and go up to,
         but may not include, the end of the video.
     5b. The ticks should be evenly spaced, and the text should
         not overlap.
     5c. The number of ticks should be maximized. Verify that
         the next smaller interval could not be used.
  6. Verify that the tick marks appropriately respond to scaling
     6a. As the scale increases, the tick intervals should
         become increasingly smaller
     6b. As the scale changes, the handle, when paused, should
         not change in visual value
     6c. The timestamps should reflect the scaled range of the
         scrubber bar
     6d. Dragging the scale bar should change the timestamps
         appropriately

Type: New Feature

* Refactors and reimplements the scrubber bar

Before this patch, the scrubber bar was a QSlider placed in a
QGraphicsScene and viewed in a QGraphicsView. Upon scaling the scrubber
bar, the view would zoom in to the respective portion of the slider bar.
While this did provide a scaling effect, it had some undesirable side
effects. The most problematic was the handle width would zoom in
addition to the bar. And, with the handle width becoming an unknown
unretrievable quantity, it has become impossible to quantitate the exact
width of the slider bar, a necessary factor in computing tick-marks.

This patch reimplements the scrubber bar as a derived QSlider, which
dynamically changes its range according to the scaling bar in order to
scale, and which hides its handle using an overriden paintEvent method.
Now, scaling should be possible, and in addition, the handle width is
constant.

This patch also refactors the scalable scrubbing bar files, moving all
related files into a related folder for better organization.

Testing Steps:
  1. Run the application
  2. Verify the scrubber bar handle can be moved within its current
     range
  3. Verify that the progress rect (the blue rectangle spanning from the
     beginning of the slider to the handle) moves with the handle
     appropriately.
  4. Load a video
  5. Verify that the scrubber bar works as expected when the scaling bar
     is set in its default state.
     5a. Verify the scrubber bar updates with the video
     5b. Verify updates in the scrubber bar update the video
  6. Change the scaling bar value and verify that the range of the
     scrubber bar accurately represents the scaling bar interval
  7. Verify that the handle is hidden when the progress bar handle
     represents a value outside the range of the scrubber bar.
     7a. When the progress bar handle is less than the minimum of the
     scrubber bar, the handle and progress rect should not be visible
     7b. When the progress bar handle is greater than the maximum of the
     scrubber bar, the handle should not be visible and the progress
     rect should occupy the entire width of the scrubber bar.

Type: New Feature

* Sets timestamp width dependent slider margins

Before this patch, the sliders that make up the scalable scrubbing bar
occupied as much vertical space as possible in the horizontal layouts
they were placed in. In order to fit timestamp labels on the extreme
ends of a slider, or in their future tick-mark bars, we need extra space
on their left and right sides. This patch adds margins, based on the
maximum width of a timestamp label, to the sides of each slider.

Testing Steps:
  1. Run the application
  2. Verify that the slider labels are left aligned
  3. Verify that each slider has the same left and right margins
  4. Verify that the timestamp labels in the scaling bar are not cut off
  5. Load a video
  6. Verify that the functionality of the sliders work as expected

Type: New Feature